pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lightspeed-agentic"
version = "0.1.0"
description = "Multi-provider agentic sandbox for OpenShift Lightspeed"
readme = "README.md"
requires-python = ">=3.12"
license = "Apache-2.0"
dependencies = ["fastapi>=0.136.0"]

[project.optional-dependencies]
claude = ["claude-agent-sdk>=0.1.66"]
gemini = ["google-adk>=1.31.1"]
openai = ["openai-agents>=0.14.5"]
deepagents = ["deepagents>=0.5.3", "langchain-google-vertexai>=3.2.2"]
all = ["lightspeed-agentic[claude,gemini,openai,deepagents]"]
dev = ["pytest>=9.0.3", "pytest-asyncio>=1.1.0", "httpx>=0.28.1"]

[tool.hatch.build.targets.wheel]
packages = ["src/lightspeed_agentic"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.12"
strict = true
